requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
flower==1.0.0
humanize==3.13.1
kombu==5.2.2
prometheus-client==0.12.0
prompt-toolkit==3.0.23
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
